Make it possible to check out this repository on Windows
On Windows, colons are not allowed in filenames, and neither are
question marks. Let's just rename the files whose filenames contain
colons or question marks (simply removing them, there are no
collisions).

The affected files (all in the ProGit book) already contain `url`
attributes in the front matter, meaning: Hugo will use the correct
filename in the rendered output, and therefore the end result is the
same as before.

This trick was performed by

  git -c core.quotePath=false ls-files |
  grep ':' |
  while read path
  do
    git mv --sparse "$path" "${path/:/}" ||
    break
  done

and

  git -c core.quotePath=false ls-files |
  grep '?' |
  while read path
  do
    git mv --sparse "$path" "${path/\?/}" ||
    break
  done

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>