test: let it pass on Windows
On Windows, file names cannot contain question marks. We describe a
work-around to allow rendering the site (which does contain pages whose
URLs contain literal question marks, for historical reasons, and that
can therefore unfortunately not be changed) by changing the affected
`url` front-matter attributes to _not_ contain question marks.

To let the resulting site pass the test suite, we need to adjust it a
little.

We also need to give the test more time because the `manual pages` test
case seems to take a bit longer than those 30 seconds which are the
default timeout of Playwright, at least in my setup.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>