README: describe the minimal setup required to pass the test suite
When trying to enhance the test suite, it can be a bit tricky to figure
out what sparse-checkout definition to use. The alternative would be to
build the entire site, which would be prohibitively slow for an
efficient development cycle.

Therefore, let's document that (quite involved, I am afraid!) sparse
checkout definition for the minimum set of files required to build the
site and run the test suite.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>