Add a print-specific style sheet
It would appear that the navigation bar is already suppressed when
printing, thanks to being inside an `<aside>` element. Other elements
that are not useful in print, such as the search box, the version
selector, the dark mode toggle, and the "Edit this page" link are
now also hidden when printing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>