Hide the dark mode button when printing
The dark mode button makes no sense in a printed document.

This change is slightly tricky because we used to hide the dark button
by default and only showed it when the JavaScript initialization ran
as expected, by setting the CSS `display` property to `block`.

However, doing that also meant that the `display` property could not be
overridden by a `@media print` block, because the `display` property was
set on the dark mode button itself, and would therefore always be more
specific than the `@media print` rule.

So let's instead use a CSS class to indicate whether the dark mode button
should be shown or not.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>