Add a GitHub workflow cache to manually purge the Cloudflare cache
We had problems with the cache-purging logic in
https://github.com/git/git-scm.com/issues/2086, and then also in
https://github.com/git/git-scm.com/issues/2093. The symptom is that the
CSS isn't loading, because the front-page still serves an older version
(one that references a no-longer-existing CSS). The culprit seems to be
the un-purged Cloudflare cache.

Let's add a way to manually purge the cache. It is totally possible that
this is one of those famous "Close Door" elevator buttons that fool us:
https://www.sciencealert.com/the-close-door-buttons-in-elevators-don-t-actually-do-anything
But in case it isn't, let's have a way to trigger a purge.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>